Add hook for Player movement (refactoring)
