+++ packages/mr-client/src/Player/Player.tsx
@@ -1,7 +1,9 @@
 import { useSphere } from '@react-three/cannon'
 import { useThree, useFrame } from '@react-three/fiber'
-import { useState, useRef, useEffect } from 'react'
+import { useRef, useEffect } from 'react'
 import { Vector3 } from 'three'
+
+import usePlayerControls from './hooks/usePlayerControls'
 
 export type PlayerProps = {
   color?: string
@@ -17,56 +19,12 @@ export default function Player({
     position: [0, 0.5, 0]
   }))
 
-  const moveFieldByKey = (key: string | number) => keys[key]
   const SPEED = 5
   const direction = new Vector3()
   const frontVector = new Vector3()
   const sideVector = new Vector3()
-  const keys = {
-    KeyW: 'forward',
-    KeyS: 'backward',
-    KeyA: 'left',
-    KeyD: 'right'
-  }
 
-  const usePlayerControls = () => {
-    const [movement, setMovement] = useState({
-      forward: false,
-      backward: false,
-      left: false,
-      right: false
-    })
-
-    useEffect(() => {
-      // TODO Handle button press down
-      const handleKeyDown = (event: { code: string | number }) => {
-        setMovement(movements => ({
-          ...movements,
-          [moveFieldByKey(event.code)]: true
-        }))
-        api.applyForce([0.5, 0.5, 0.5], [0, 0, 0])
-      }
-
-      // TODO Handle button press up
-      const handleKeyUp = (event: { code: string | number }) => {
-        setMovement(movements => ({
-          ...movements,
-          [moveFieldByKey(event.code)]: false
-        }))
-      }
-
-      document.addEventListener('keydown', handleKeyDown)
-      document.addEventListener('keyup', handleKeyUp)
-      return () => {
-        document.removeEventListener('keydown', handleKeyDown)
-        document.removeEventListener('keyup', handleKeyUp)
-      }
-    }, [])
-
-    return movement
-  }
-
-  const { forward, backward, left, right } = usePlayerControls()
+  const { forward, backward, left, right } = usePlayerControls(api)
   const { camera } = useThree()
   const velocity = useRef([0, 0, 0])
